move marginal_model to model/
